reorganize `util` mod.
