+++ embassy/src/util/steal.rs
@@ -1,0 +1,3 @@
+pub trait Steal {
+    unsafe fn steal() -> Self;
+}
